Created methods to better analyze operations performed on speed array
